+++ src/org/starloco/locos/common/PathFinding.java
@@ -211,7 +211,7 @@ public class PathFinding {
     }
 
 
-    public final static Map<String, List<Short>> outForbiddenCells = new HashMap<>();
+    public final static Map<String, List<Integer>> outForbiddenCells = new HashMap<>();
 
     public static int GetCaseIDFromDirrection(int cellId, char dir,
                                               GameMap map, boolean fight) {
@@ -244,8 +244,8 @@ public class PathFinding {
                 cell =  cellId - map.getW() + 1;
                 break;
         }
-        List<Short> cells = outForbiddenCells.get(map.getW() + "_" + map.getH());
-        if(cells != null && cells.contains((short) cell))
+        List<Integer> cells = outForbiddenCells.get(map.getW() + "_" + map.getH());
+        if(cells != null && cells.contains(cell))
             return -1;
         return cell;
     }
@@ -1175,7 +1175,7 @@ public class PathFinding {
             return -1;
         if (forbidden == null) forbidden = new ArrayList<>();
         int dist = 1000, cellId = startCell;
-        List<Short> cells = outForbiddenCells.get(map.getW() + "_" + map.getH());
+        List<Integer> cells = outForbiddenCells.get(map.getW() + "_" + map.getH());
 
         for (char d : new char[] {'b', 'd', 'f', 'h'}) {
             int newCellId = PathFinding.GetCaseIDFromDirrection(startCell, d, map, true);
@@ -1185,7 +1185,7 @@ public class PathFinding {
                 int distance = PathFinding.getDistanceBetween(map, endCell, newCellId);
 
                 if (distance < dist && cell.isWalkable(true, true, -1) && cell.getFirstFighter() == null
-                        && !forbidden.contains(cell) && !cells.contains((short) newCellId)) {
+                        && !forbidden.contains(cell) && !cells.contains(newCellId)) {
                     dist = distance;
                     cellId = newCellId;
                 }
